+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * On Linux, the `rbw-agent` process can no longer be attached to by debuggers,
   and no longer produces core dumps (#42, oranenj)
+* Suggest rotating the user's encryption key if we see an old cipherstring type
+  (#40, rjc)
 
 ## [1.0.0] - 2021-02-21